Replaced YAML ast with custom parser.
This should be several commits due to changing how aliases and patterns are implemented, but back-porting these changes to the yaml ast to just throw it away is wasteful, so it's done all at once.
We will also be simplified a large part of the codebase in the near time, so many of these changes are patchwork to make things work.
Due to this, the guide also needs to be rewritten from scratch, and so was currently removed and will be added again eventually.